refactor to OPI_Utilities
