+++ SKO_Hub/SKO_HubClient.h
@@ -9,8 +9,8 @@
 
 
 #include <thread>
-#include "../SKO_Utilities/OPI_Clock.h"
-#include "../SKO_Utilities/OPI_Sleep.h"
+#include "../OPI_Utilities/OPI_Clock.h"
+#include "../OPI_Utilities/OPI_Sleep.h"
 #include <mutex>
 
 #include <stdio.h>
