+++ SKO_Network/SKO_ChatHandler.cpp
@@ -1,5 +1,5 @@
 #include "SKO_ChatHandler.h"
-#include "../SKO_Utilities/SKO_Utilities.h"
+#include "../OPI_Utilities/OPI_Utilities.h"
 
 
 SKO_ChatHandler::SKO_ChatHandler(SKO_Network *network)
@@ -10,7 +10,7 @@ SKO_ChatHandler::SKO_ChatHandler(SKO_Network *network)
 // [CHAT]["/ban"][" "]["<username>"][" "]["<reason>"]
 void SKO_ChatHandler::parseSlashBan(unsigned char userId, std::string parameters)
 {
-    std::string username = SKO_Utilities::nextParameter(parameters);
+    std::string username = OPI_Utilities::nextParameter(parameters);
     std::string reason = parameters;
     network->banPlayer(userId, username, reason);
 }
@@ -27,7 +27,7 @@ void SKO_ChatHandler::parseSlashUnban(unsigned char userId, std::string paramete
 void SKO_ChatHandler::parseSlashMute(unsigned char userId, std::string parameters)
 {
     //strip the appropriate data
-    std::string username = SKO_Utilities::nextParameter(parameters);
+    std::string username = OPI_Utilities::nextParameter(parameters);
     std::string reason = parameters;
     network->mutePlayer(userId, username, reason);
 }
@@ -35,14 +35,14 @@ void SKO_ChatHandler::parseSlashMute(unsigned char userId, std::string parameter
 // [CHAT]["/unmute"][" "]["<username>"]
 void SKO_ChatHandler::parseSlashUnmute(unsigned char userId, std::string parameters)
 {
-    std::string username = SKO_Utilities::nextParameter(parameters);
+    std::string username = OPI_Utilities::nextParameter(parameters);
     network->unmutePlayer(userId, username);
 }
 
 // [CHAT]["/ban"][" "]["<username>"][" "]["<reason>"]
 void SKO_ChatHandler::parseSlashKick(unsigned char userId, std::string parameters)
 {
-    std::string username = SKO_Utilities::nextParameter(parameters);
+    std::string username = OPI_Utilities::nextParameter(parameters);
     std::string reason = parameters;
     network->kickPlayer(userId, username, reason);
 }
@@ -57,7 +57,7 @@ void SKO_ChatHandler::parseSlashWho(unsigned char userId)
 // [CHAT]["/ipban"][" "]["<ipv4 address>"][" "][<reason>]
 void SKO_ChatHandler::parseSlashIpban(unsigned char userId, std::string parameters)
 {
-    std::string ip = SKO_Utilities::nextParameter(parameters);
+    std::string ip = OPI_Utilities::nextParameter(parameters);
     std::string reason = parameters;
     network->banAddress(userId, ip, reason);
 }
@@ -65,23 +65,23 @@ void SKO_ChatHandler::parseSlashIpban(unsigned char userId, std::string paramete
 // [CHAT]["/getip"][" "]["<player>"]
 void SKO_ChatHandler::parseSlashGetip(unsigned char userId, std::string parameters)
 {
-    std::string username = SKO_Utilities::nextParameter(parameters);
+    std::string username = OPI_Utilities::nextParameter(parameters);
     network->getIp(userId, username);
 }
 // [CHAT]["/warp"][" "]["<player>"][" "][(string)x][" "][(string)y][" "][(string)mapId]
 void SKO_ChatHandler::parseSlashWarp(unsigned char userId, std::string parameters)
 {
-    std::string username = SKO_Utilities::nextParameter(parameters);
-    int x = atoi(SKO_Utilities::nextParameter(parameters).c_str());
-    int y = atoi(SKO_Utilities::nextParameter(parameters).c_str());
-    int mapId = atoi(SKO_Utilities::nextParameter(parameters).c_str());
+    std::string username = OPI_Utilities::nextParameter(parameters);
+    int x = atoi(OPI_Utilities::nextParameter(parameters).c_str());
+    int y = atoi(OPI_Utilities::nextParameter(parameters).c_str());
+    int mapId = atoi(OPI_Utilities::nextParameter(parameters).c_str());
     network->warpPlayer(userId, username, x, y, mapId);                
 }
 
 // [CHAT]["/ping"]
 void SKO_ChatHandler::parseSlashPing(unsigned char userId, std::string parameters)
 {
-    std::string username = SKO_Utilities::nextParameter(parameters);
+    std::string username = OPI_Utilities::nextParameter(parameters);
     network->showPlayerPing(userId, username);
 }
