+++ SKO_Network/SKO_PacketHandler.cpp
@@ -1,7 +1,7 @@
 #include "SKO_PacketHandler.h"
 #include "SKO_ChatHandler.h"
 #include "SKO_Network.h"
-#include "../SKO_Utilities/SKO_Utilities.h"
+#include "../OPI_Utilities/OPI_Utilities.h"
 #include "../Global.h"
 
 //constructor
@@ -30,7 +30,7 @@ void SKO_PacketHandler::parseLogin(unsigned char userId, SKO_PacketParser *parse
 {
     // Declare message string
     std::string loginRequest = parser->getPacketBody();
-    std::string username = SKO_Utilities::nextParameter(loginRequest);
+    std::string username = OPI_Utilities::nextParameter(loginRequest);
     std::string password = loginRequest;
     network->attemptLogin(userId, username, password);
 }
@@ -38,7 +38,7 @@ void SKO_PacketHandler::parseLogin(unsigned char userId, SKO_PacketParser *parse
 // [REGISTER][<username>][" "][<password>]
 void SKO_PacketHandler::parseRegister(unsigned char userId, std::string parameters)
 {
-    std::string username = SKO_Utilities::nextParameter(parameters);
+    std::string username = OPI_Utilities::nextParameter(parameters);
     std::string password = parameters;//do not use nextParameter to get password, just use remaining
     network->attemptRegister(userId, username, password);      
 }
